bash PS1 showing git verbs
This does more than git status and could be broken into just the git
part if that was interesting.  The git part prints verbs in English
suggesting next steps in normal workflow, so the prompt could be
something like "(add master)$ " if we have an untracked file that could
be added, or "(commit master)$ " if we have a staged change that isn't
in the repo yet.

Also does bash magic to print a pretty two-line prompt which depends on
my pwd_short script.  Okay that part should be broken out.