catch git_prompt stderr as bad output
Test should match stderr as well as stdout so we can report errors as
failures.  Also strengthened the regex check to be required to match the
whole line (grep -x) because we had "master" in the error message and it
was passing by mistake.

Also bust git_cautious_check_remote_cache so we test reading upstream
every run.

status quo: git_prompt is broken with bad upstream