improved cache key
Keeping the human readable part, the shasum should be on the origin url
rather than the repo path.  This way if multiple repos share the same
project name and same origin, they share the cached check.  This also
means that when the origin is changed then cache invalidates and we'll
immediately go off to look at the new origin.