async network steps
If we have an answer cached, use it, but when we have to go look at the
remote over the network, do that in a subshell and return "no updates".
The main impact of this is on git_prompt, which will be snappy fast but
still eventually pick up the need for a pull.